Rename bmp280 to bmp280_i2c.
This is in preparation for adding a bmp280_spi component, similar to
