--- tests/components/bmp280/test.esp32-c3-idf.yaml
@@ -1,15 +1,0 @@
-i2c:
-  - id: i2c_bmp280
-    scl: 5
-    sda: 4
-
-sensor:
-  - platform: bmp280
-    address: 0x77
-    temperature:
-      name: Outside Temperature
-      oversampling: 16x
-    pressure:
-      name: Outside Pressure
-    iir_filter: 16x
-    update_interval: 15s
