+++ tests/components/bmp280_i2c/test.esp8266-ard.yaml
@@ -1,0 +1,5 @@
+substitutions:
+  scl_pin: GPIO5
+  sda_pin: GPIO4
+
+<<: !include common.yaml
